# docs/swagger.yaml
basePath: /v1
host: localhost:8080
info:
  contact:
    email: nguyenminhmannh2001@gmail.com
    name: Nguyen Minh Manh
  description: A centralized authentication hub for my home applications in Go using
    Gin framework.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: AuthConnect Hub
  version: "1.0"
paths:
...
  /v1/auth/login:
    get:
      description: This endpoint renders the login page and displays a toast notification
        if provided query parameters are valid.
      parameters:
      - description: The message to display in the toast notification.
        in: query
        name: toast-message
        type: string
      - description: The type of the toast notification (e.g., success, error).
        in: query
        name: toast-type
        type: string
      - description: A hash value used for validation.
        in: query
        name: hash-value
        type: string
      produces:
      - text/html
      responses: {}
      summary: Login Page
      tags:
      - Authen
    post:
      description: Logs in a user with the provided credentials.
      produces:
      - text/html
      responses: {}
      summary: User Login
      tags:
      - Authen
  /v1/auth/logout:
    get:
      description: Logs out the currently authenticated user and redirects to the
        home page with a success toast notification.
      responses: {}
      security:
      - JWT: []
      summary: Logout User
      tags:
      - Authen
...
  /v1/dashboard:
    get:
      description: Retrieves the dashboard data for the currently authenticated user.
      produces:
      - text/html
      responses: {}
      security:
      - JWT: []
      summary: Get Dashboard Data
      tags:
      - Dashboard
  /v1/dashboard/profile:
    get:
      description: Retrieves the profile information for the currently authenticated
        user.
      parameters:
      - description: The message to display in the toast notification.
        in: query
        name: toast-message
        type: string
      - description: The type of the toast notification (e.g., success, error).
        in: query
        name: toast-type
        type: string
      - description: A hash value used for validation.
        in: query
        name: hash-value
        type: string
      produces:
      - text/html
      responses: {}
      security:
      - JWT: []
      summary: Get User Profile
      tags:
      - Dashboard
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"